requirements.txt
apache-airflow[snowflake]==2.6.1
dbt-snowflake
snowflake-connector-python==2.4.5